* Added menu item "Edit->File encoding" to select the active file's encoding. Partial patch by 280Z28. * Added simple BOM signature encoding auto-detection on file open. * Convert filenames on open. This allows to open the same project file in all supported OSes. * Updated AngelScript to 2.5.0b. * Fixed all warnings when building with gcc-4.x.
git-svn-id: https://svn.code.sf.net/p/codeblocks/code/trunk@1943 2a5c6006-c6dd-42ca-98ab-0921f2732cef

+++ src/sdk/as/source/as_callfunc_x86.cpp
@@ -456,7 +456,7 @@ int CallSystemFunction(int id, asCContext *context, void *objectPointer)
 		    args++;
 
 		int spos = 0;
-		for( size_t n = 0; n < descr->parameterTypes.GetLength(); n++ )
+		for( int n = 0; n < (int)descr->parameterTypes.GetLength(); n++ )
 		{
 			if( descr->parameterTypes[n].IsObject() &&
 				!descr->parameterTypes[n].IsReference() &&
